requirement.txt
aioredis==1.3.1
anyio==3.6.2
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
autobahn==23.1.1
Automat==22.10.0
black==22.12.0
certifi==2022.12.7
cffi==1.15.1
channels==3.0.5
channels-redis==3.4.1
charset-normalizer==3.0.1
click==8.1.3
constantly==15.1.0
cryptography==39.0.0
daphne==3.0.1
datetime-truncate==1.1.1
elasticsearch==7.17.9
elasticsearch-dsl==7.4.1
Django==4.0.5
django-admin-charts==1.2.4
django-bower==5.2.0
django-cors-headers==3.13.0
django-environ==0.9.0
django-memoize==2.3.1
django-multiselectfield==0.1.12
django-nvd3==0.9.7
django-rest-auth-dj4==0.0.6
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.5
filetype==1.2.0
h11==0.14.0
hiredis==2.2.1
httptools==0.5.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
Jinja2==3.1.2
kafka-python==2.0.2
MarkupSafe==2.1.1
msgpack==1.0.4
mypy-extensions==0.4.3
pathspec==0.10.3
Pillow==9.4.0
platformdirs==2.6.2
psycopg2-binary==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==1.7.1
pyOpenSSL==23.0.0
python-dateutil==2.8.2
python-dotenv==0.21.0
python-nvd3==0.14.2
python-slugify==1.1.4
pytz==2022.7
PyYAML==6.0
redis==4.4.2
requests==2.28.2
service-identity==21.1.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.3
tomli==2.0.1
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.4.0
Unidecode==1.3.6
urllib3==1.26.14
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
zope.interface==5.5.2
elasticsearch==7.17.9
elasticsearch-dsl==7.4.1
pytest==7.3.1
pytest_django==4.5.2
pytest_cov==4.0.0
psycopg_binary==3.1.9
djangorestframework_gis==1.0.0
